[Fix #11782] Move pending cops warning out of ConfigLoader
Previously, `RuboCop::ConfigLoader` emitted pending cops warnings as a side effect of loading configuration files.
This caused unintended stdout output during tests and could interfere with CLI operations like `--auto-gen-config`.

This patch moves the responsibility for pending cops reporting to a dedicated `PendingCopsReporter` class.
The CLI now explicitly invokes the reporter after loading configuration, ensuring that `ConfigLoader` remains a pure loader without user-facing output.
