+++ changelog/fix_move_pending_cops_warning_out_of_config_loader_20250703225645.md
@@ -1,0 +1,1 @@
+* [#11782](https://github.com/rubocop/rubocop/issues/11782): Move pending cops warning out of ConfigLoader. ([@nobuyo][])
