+++ spec/support/suppress_pending_warning.rb
@@ -2,11 +2,11 @@
 
 #
 # This is a file that supports testing. An open class has been applied to
-# `RuboCop::ConfigLoader.warn_on_pending_cops` to suppress pending cop
+# `RuboCop::PendingCopsReporter.warn_on_pending_cops` to suppress pending cop
 # warnings during testing.
 #
 module RuboCop
-  class ConfigLoader
+  class PendingCopsReporter
     class << self
       remove_method :warn_on_pending_cops
       def warn_on_pending_cops(config)
